Added onReads
The Angle and Status Characteristics now have onRead callbacks. When read, they will update their values

+++ WheelchairSensor/IMUSensors.h
@@ -49,7 +49,7 @@ struct IMUSensor {
   float pitch;
 
   // Connected status and the IMU object
-  int connected;
+  uint8_t connected;
   uint8_t id;
   Adafruit_LSM6DSO32 imu;
 } ;
